requirements.txt
pandas
pyyaml
requests
fastapi
uvicorn
python-dotenv
